Tercera práctica integradora

+++ src/controllers/products.controllers.js
@@ -58,6 +58,7 @@ export const addProduct = async (req,res)=>{
          error:customerError,
        })
    }
+   product.owner = req.user._id;
    const resultado = await productManager.addProduct(product)
    if(resultado){
       return res.send({
@@ -78,12 +79,18 @@ export const updateProduct = async (req,res)=>{
 }
 
 export const deleteProduct = async (req,res)=>{
-   const id = req.params.pid;   
-   const resultado = await productManager.deleteProduct(id)
-   if(resultado){
-      return res.send({
-         producto:resultado,
-      })
+   const id = req.params.pid;
+   const productOwer = JSON.parse(JSON.stringify(product.owner));
+   const userId = JSON.parse(JSON.stringify(req.user._id));
+   if((req.user.rol === "premium" && productOwer == userId) || req.user.rol === "admin"){
+      const resultado = await productManager.deleteProduct(id)
+      if(resultado){
+         return res.send({
+            producto:resultado,
+         })
+      }
+   }else{
+      res.send({status:"error", message:"no puedes borrar este producto"})
    }
 }
